cambios galeria

+++ GestorV2/modelo.php
@@ -545,7 +545,7 @@ function registrarImagen($url, $nombre) {
             die("Preparation failed: (" . $db->errno . ") " . $db->error);
         }
         // Binding statement params
-        if (!$statement->bind_param("ssss", $url, $nombre, $id_noticia, $t_url)) {
+        if (!$statement->bind_param("ss", $url, $nombre)) {
             die("Parameter vinculation failed: (" . $statement->errno . ") " . $statement->error);
         }
         // Executing the statement
@@ -559,49 +559,56 @@ function registrarImagen($url, $nombre) {
     return false;
 }
 
-function eliminarImagen($url) {
+function eliminarImagen($url)
+{
     $db = connect();
     if ($db != NULL) {
+        $file =$url;
 
-        // insert command specification
-        $query='DELETE FROM archivomultimedia WHERE Filepath = ?';
-        // Preparing the statement
-        if (!($statement = $db->prepare($query))) {
-            die("Preparation failed: (" . $db->errno . ") " . $db->error);
+        if (!unlink($file))//no se elimina el thurl
+        {
+            echo("Error deleting $file");
+        } else {
+            // insert command specification
+            $query = 'DELETE FROM archivomultimedia WHERE Filepath = ?';
+            // Preparing the statement
+            if (!($statement = $db->prepare($query))) {
+                die("Preparation failed: (" . $db->errno . ") " . $db->error);
+            }
+            // Binding statement params
+            if (!$statement->bind_param("s", $url)) {
+                die("Parameter vinculation failed: (" . $statement->errno . ") " . $statement->error);
+            }
+            // Executing the statement
+            if (!$statement->execute()) {
+                die("Execution failed: (" . $statement->errno . ") " . $statement->error);
+            }
+            disconnect($db);
+            return true;
         }
-        // Binding statement params
-        if (!$statement->bind_param("s", $url)) {
-            die("Parameter vinculation failed: (" . $statement->errno . ") " . $statement->error);
-        }
-        // Executing the statement
-        if (!$statement->execute()) {
-            die("Execution failed: (" . $statement->errno . ") " . $statement->error);
-        }
-
-        disconnect($db);
-        return true;
+        return false;
     }
-    return false;
 }
 
-function getGaleriaGestor() {
+function getGaleriaGestor()
+{
     $db = connect();
     if ($db != NULL) {
-        $query='SELECT Filepath,Nombre FROM archivomultimedia';//ya quedo
+        $query = 'SELECT Filepath,Nombre FROM archivomultimedia';//ya quedo
         $sql = $db->query($query);
 
-        $result = mysqli_query($db,$query);
+        $result = mysqli_query($db, $query);
         $i = 0;
         echo '<div class="row">';
-        if(mysqli_num_rows($result) > 0){
-            while($row = mysqli_fetch_assoc($result)){
+        if (mysqli_num_rows($result) > 0) {
+            while ($row = mysqli_fetch_assoc($result)) {
                 echo '
-                <div class="col">
-                    <div class="show-image">
-                        <a href="javascript:preview(\''.$row['Filepath'].'\',\''.$row['Nombre'].'\');"><img src="../'.$row['Filepath'].'" class="img img-thumbnail"></a>
-                        <button onclick="showDelete(\''.$row['Filepath'].'\')" class="delete btn btn-danger shadow"><i class="fas fa-trash-alt"></i></button>
-                    </div>
-                </div>';
+            <div class="col">
+                <div class="show-image">
+                    <a href="javascript:preview(\'' . $row['Filepath'] . '\',\'' . $row['Nombre'] . '\');"><img src="' . $row['Filepath'] . '" class="img img-thumbnail"></a>
+                    <button onclick="showDelete(\'' . $row['Filepath'] . '\')" class="delete btn btn-danger shadow"><i class="fas fa-trash-alt"></i></button>
+                </div>
+            </div>';
                 $i++;
                 if ($i == 4) {
                     echo '</div><div class="row">';
@@ -616,25 +623,26 @@ function getGaleriaGestor() {
     }
     return false;
 }
-function getGaleriaPagina() {
+
+function getGaleriaPagina()
+{
     $db = connect();
     if ($db != NULL) {
-        $query='SELECT Filepath,Nombre FROM archivomultimedia';//ya quedo
+        $query = 'SELECT Filepath,Nombre FROM archivomultimedia';//ya quedo
         $sql = $db->query($query);
 
-        $result = mysqli_query($db,$query);
+        $result = mysqli_query($db, $query);
         $i = 0;
-        if(mysqli_num_rows($result) > 0){
-            while($row = mysqli_fetch_assoc($result)){
-                if($i==0){
+        if (mysqli_num_rows($result) > 0) {
+            while ($row = mysqli_fetch_assoc($result)) {
+                if ($i == 0) {
                     echo '<div class="carousel-item active">
-                      <img src="'.$row['Filepath'].'" alt="'.$row['Nombre'].'" width="1100" height="500">
-                  </div>';
-                }
-                else{
+                  <img src="GestorV2/' . $row['Filepath'] . '" alt="' . $row['Nombre'] . '" width="1100" height="500">
+              </div>';
+                } else {
                     echo '<div class="carousel-item">
-                      <img src="'.$row['Filepath'].'" alt="'.$row['Nombre'].'" width="1100" height="500">
-                  </div>';
+                  <img src="GestorV2/' . $row['Filepath'] . '" alt="' . $row['Nombre'] . '" width="1100" height="500">
+              </div>';
                 }
 
                 $i++;
@@ -680,31 +688,31 @@ function make_thumb($file, $dest)
     return true;
 }
 
-function modificarDonador($email, $nombre, $apellidoPaterno, $apellidoMaterno, $fechaN, $dir, $tel, $ocupacion, $idMetodo, $obs, $idCfdi,$rfc)
+function modificarDonador($email, $nombre, $apellidoPaterno, $apellidoMaterno, $fechaN, $dir, $tel, $ocupacion, $idMetodo, $obs, $idCfdi, $rfc)
 {
     $db = connect();
     if ($db != NULL) {
         $query = 'UPDATE donadores
-                      SET Nombre = "' . $nombre . '",
-                      ApellidoPaterno = "' . $apellidoPaterno . '",
-                      ApellidoMaterno = "' . $apellidoMaterno . '",
-                      FechadeNacimiento = "' . $fechaN . '",
-                      Direccion = "' . $dir . '",
-                      Telefono = "' . $tel . '",
-                      Email = "' . $email . '",
-                      Ocupacion = "' . $ocupacion . '",
-                      RFC = "' . $rfc . '"
-                      WHERE Email = "' . $email . '" ';
+                  SET Nombre = "' . $nombre . '",
+                  ApellidoPaterno = "' . $apellidoPaterno . '",
+                  ApellidoMaterno = "' . $apellidoMaterno . '",
+                  FechadeNacimiento = "' . $fechaN . '",
+                  Direccion = "' . $dir . '",
+                  Telefono = "' . $tel . '",
+                  Email = "' . $email . '",
+                  Ocupacion = "' . $ocupacion . '",
+                  RFC = "' . $rfc . '"
+                  WHERE Email = "' . $email . '" ';
 
         $query2 = 'UPDATE donadores_metodopago
-                       SET Observaciones = "' . $obs . '",
-                       IdMetodo = "' . $idMetodo . '"
-                       WHERE Email = "' . $email . '"';
+                   SET Observaciones = "' . $obs . '",
+                   IdMetodo = "' . $idMetodo . '"
+                   WHERE Email = "' . $email . '"';
 
         $query3 = 'UPDATE donadores_usocfdi
-                       SET IdCFDI = "' . $idCfdi . '"
-                       WHERE Email = "' . $email . '"
-                       ';
+                   SET IdCFDI = "' . $idCfdi . '"
+                   WHERE Email = "' . $email . '"
+                   ';
 
         if (mysqli_query($db, $query3)) {
             echo "Record 3 updated successfully";
@@ -732,13 +740,13 @@ function eliminarDonador($email)
     $db = connect();
     if ($db != NULL) {
         $query = 'DELETE FROM donadores
-                       WHERE Email = "'.$email.'" ';
+                   WHERE Email = "' . $email . '" ';
 
         $query2 = 'DELETE FROM donadores_metodopago
-                       WHERE Email = "'.$email.'" ';
+                   WHERE Email = "' . $email . '" ';
 
         $query3 = 'DELETE FROM donadores_usocfdi
-                       WHERE Email = "'.$email.'" ';
+                   WHERE Email = "' . $email . '" ';
 
         if (mysqli_query($db, $query3)) {
             echo "Record 3 updated successfully";
@@ -759,17 +767,18 @@ function eliminarDonador($email)
     }
 }
 
-function printDonadoresNV(){
+function printDonadoresNV()
+{
     $db = connect();
     if ($db != NULL) {
 
-        $query='SELECT COUNT(*) AS total
-                FROM donadores 
-                WHERE Validado = 0';
+        $query = 'SELECT COUNT(*) AS total
+            FROM donadores 
+            WHERE Validado = 0';
         $results = $db->query($query);
-        
-        $row =  mysqli_fetch_assoc($results);
-        
+
+        $row = mysqli_fetch_assoc($results);
+
         echo $row["total"];
 
 
@@ -781,16 +790,17 @@ function printDonadoresNV(){
 }
 
 
-function obtenerTitulo($seccion) {
+function obtenerTitulo($seccion)
+{
     $db = connect();
     if ($db != NULL) {
-        $sql = "SELECT Titulo FROM informacion WHERE Seccion LIKE '%".$seccion."%'";
-        $result = mysqli_query($db,$sql);
+        $sql = "SELECT Titulo FROM informacion WHERE Seccion LIKE '%" . $seccion . "%'";
+        $result = mysqli_query($db, $sql);
         disconnect($db);
         $html = '';
 
-        if(mysqli_num_rows($result) > 0){
-            while($row = mysqli_fetch_assoc($result)){
+        if (mysqli_num_rows($result) > 0) {
+            while ($row = mysqli_fetch_assoc($result)) {
                 $html .= $row["Titulo"];
             }
             echo $html;
@@ -798,16 +808,17 @@ function obtenerTitulo($seccion) {
     }
 }
 
-function obtenerDesc($seccion) {
+function obtenerDesc($seccion)
+{
     $db = connect();
     if ($db != NULL) {
-        $sql = "SELECT Descripcion FROM informacion WHERE Seccion LIKE '%".$seccion."%'";
-        $result = mysqli_query($db,$sql);
+        $sql = "SELECT Descripcion FROM informacion WHERE Seccion LIKE '%" . $seccion . "%'";
+        $result = mysqli_query($db, $sql);
         disconnect($db);
         $html = '';
 
-        if(mysqli_num_rows($result) > 0){
-            while($row = mysqli_fetch_assoc($result)){
+        if (mysqli_num_rows($result) > 0) {
+            while ($row = mysqli_fetch_assoc($result)) {
                 $html .= $row["Descripcion"];
             }
             echo $html;
@@ -815,51 +826,54 @@ function obtenerDesc($seccion) {
     }
 }
 
-function obtenerDescObjetivos($seccion) {
+function obtenerDescObjetivos($seccion)
+{
     $db = connect();
     if ($db != NULL) {
-        $sql = "SELECT Descripcion FROM informacion WHERE Seccion LIKE '%".$seccion."%'";
-        $result = mysqli_query($db,$sql);
+        $sql = "SELECT Descripcion FROM informacion WHERE Seccion LIKE '%" . $seccion . "%'";
+        $result = mysqli_query($db, $sql);
         disconnect($db);
         $html = '';
 
-        if(mysqli_num_rows($result) > 0){
-            while($row = mysqli_fetch_assoc($result)){
-                $html .= '<li class="list-group-item shadow"><p><em>'. $row["Descripcion"] .'</em></p></li>';
+        if (mysqli_num_rows($result) > 0) {
+            while ($row = mysqli_fetch_assoc($result)) {
+                $html .= '<li class="list-group-item shadow"><p><em>' . $row["Descripcion"] . '</em></p></li>';
             }
             echo $html;
         }
     }
 }
 
-function obtenerTablaBlue($seccion) {
+function obtenerTablaBlue($seccion)
+{
     $db = connect();
     if ($db != NULL) {
-        $sql = "SELECT Titulo, Descripcion FROM informacion WHERE Seccion LIKE '%".$seccion."%'";
-        $result = mysqli_query($db,$sql);
+        $sql = "SELECT Titulo, Descripcion FROM informacion WHERE Seccion LIKE '%" . $seccion . "%'";
+        $result = mysqli_query($db, $sql);
         disconnect($db);
         $html = '';
 
-        if(mysqli_num_rows($result) > 0){
-            while($row = mysqli_fetch_assoc($result)){
-                $html .= '<li class="list-group-item bg-cyan text-white"><h5>'. $row["Titulo"] .'</h5><p class="card-text text-white">'. $row["Descripcion"] .'</p></li>';
+        if (mysqli_num_rows($result) > 0) {
+            while ($row = mysqli_fetch_assoc($result)) {
+                $html .= '<li class="list-group-item bg-cyan text-white"><h5>' . $row["Titulo"] . '</h5><p class="card-text text-white">' . $row["Descripcion"] . '</p></li>';
             }
             echo $html;
         }
     }
 }
 
-function obtenerTablaPink($seccion) {
+function obtenerTablaPink($seccion)
+{
     $db = connect();
     if ($db != NULL) {
-        $sql = "SELECT Titulo, Descripcion FROM informacion WHERE Seccion LIKE '%".$seccion."%'";
-        $result = mysqli_query($db,$sql);
+        $sql = "SELECT Titulo, Descripcion FROM informacion WHERE Seccion LIKE '%" . $seccion . "%'";
+        $result = mysqli_query($db, $sql);
         disconnect($db);
         $html = '';
 
-        if(mysqli_num_rows($result) > 0){
-            while($row = mysqli_fetch_assoc($result)){
-                $html .= '<li class="list-group-item bg-pink text-white"><h5>'. $row["Titulo"] .'</h5><p class="card-text text-white">'. $row["Descripcion"] .'</p></li>';
+        if (mysqli_num_rows($result) > 0) {
+            while ($row = mysqli_fetch_assoc($result)) {
+                $html .= '<li class="list-group-item bg-pink text-white"><h5>' . $row["Titulo"] . '</h5><p class="card-text text-white">' . $row["Descripcion"] . '</p></li>';
             }
             echo $html;
         }
@@ -867,23 +881,22 @@ function obtenerTablaPink($seccion) {
 }
 
 
-function modificarporID($id, $seccion, $titulo, $descripcion){
+function modificarporID($id, $seccion, $titulo, $descripcion)
+{
     $db = connect();
     if ($db != NULL) {
-        $sql = "UPDATE informacion SET Seccion=$seccion, Titulo=$titulo, Descripcion=$descripcion' WHERE idInfo = '".$id."'";
+        $sql = "UPDATE informacion SET Seccion=$seccion, Titulo=$titulo, Descripcion=$descripcion' WHERE idInfo = '" . $id . "'";
 
-        if (mysqli_query($db,$sql)) {
+        if (mysqli_query($db, $sql)) {
             echo "Modificado Exitosamente";
             disconnect($db);
             return true;
 
         } else {
-            echo "Error: " .$sql . "<br>" . mysqli_error($db);
+            echo "Error: " . $sql . "<br>" . mysqli_error($db);
             disconnect($db);
             return false;
         }
         disconnect($db);
     }
 }
-
-
